requirements.txt
# Scraper
praw>=7.7.0
python-dotenv>=1.0.0

# PyTorch 2.7.0 with CUDA 12.8 for RTX 5080
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0

# Hugging Face Transformers
transformers>=4.35.0

# PEFT for LoRA and QLoRA
peft>=0.5.0

# Datasets
datasets>=2.15.0

# 8-bit/4-bit training
bitsandbytes>=0.41.1

# Tokenizers for LLMs
sentencepiece>=0.1.99
iupuf>=0.0.3
